requirements.txt
absl-py==2.1.0
accelerate==0.34.2
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
beartype==0.18.5
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cn2an==0.5.20
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.1
decorator==4.4.2
dill==0.3.8
Distance==0.1.3
dnspython==2.7.0
e2-tts-pytorch==1.0.5
einops==0.8.0
einx==0.3.0
ema-pytorch==0.6.2
email_validator==2.2.0
encodec==0.1.1
exceptiongroup==1.2.2
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.54.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.3.1
g2p-en==2.1.0
g2p_zh_en==0.1.1
gradio==4.37.2
gradio_client==1.0.2
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_resources==6.4.5
inflect==7.0.0
jaxtyping==0.2.34
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.9.2
llvmlite==0.43.0
loguru==0.7.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.10.0
networkx==3.3
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.29.0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
proces==0.1.7
proglog==0.1.10
progressbar33==2.4
protobuf==5.28.2
psutil==6.0.0
pyarrow==17.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
pypinyin==0.44.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
resampy==0.4.3
rich==13.9.3
rpds-py==0.20.0
ruff==0.7.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
starlette==0.37.2
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
timm==1.0.11
tokenizers==0.20.1
tomlkit==0.12.0
torch==2.4.1
torchaudio==2.4.1
torchdiffeq==0.2.4
torchlibrosa==0.1.0
torchvision==0.19.1
tqdm==4.66.5
transformers==4.46.0
triton==3.0.0
typeguard==2.13.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
vocos==0.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.4
x-transformers==1.37.4
xxhash==3.5.0
yarl==1.12.1
